WIP: moved serialization/deserialization into same shared file
