+++ packages/qwik/src/serializer/index.unit.tsx
@@ -6,15 +6,16 @@ import { isJSXNode } from '../core/render/jsx/jsx-runtime';
 import { getContainer, processVNodeData } from './client/api';
 import type { Container, VNode } from './client/types';
 import type { Stringifiable } from './shared-types';
-import { SerializationConstant, isStringifiable } from './shared-types';
+import { isStringifiable } from './shared-types';
 import { SsrNode, ssrCreateContainer, toSsrAttrs } from './ssr/api';
 import './vdom-diff.unit';
 import { vnode_getFirstChild } from './client/vnode';
-import { isDeserializerProxy } from './client/deserializer';
+import { isDeserializerProxy } from './shared-serialialization';
 import { component$ } from '../core/component/component.public';
 import { inlinedQrl, qrl } from '../core/qrl/qrl';
 import type { QRLInternal } from '../core/qrl/qrl-class';
 import { SERIALIZABLE_STATE } from '../core/container/serializers';
+import type { SSRContainer } from './ssr/types';
 
 describe('serializer v2', () => {
   describe('basic use cases', () => {
@@ -74,8 +75,8 @@ describe('serializer v2', () => {
     it('should serialize object', () => {
       const container = withContainer((ssrContainer) => {
         const obj = { age: 1, child: { b: 'child' } };
-        expect(ssrContainer.getObjectId(obj)).toBe(0);
-        expect(ssrContainer.getObjectId(obj.child)).toBe(1);
+        expect(ssrContainer.addRoot(obj)).toBe(0);
+        expect(ssrContainer.addRoot(obj.child)).toBe(1);
       });
       const obj = container.getObjectById(0);
       expect(obj).toEqual({ age: 1, child: { b: 'child' } });
@@ -84,7 +85,7 @@ describe('serializer v2', () => {
 
     it('should escape string <>', () => {
       const container = withContainer((ssrContainer) => {
-        ssrContainer.getObjectId({ '<scrip></script>': '"<scrip></script>"', '<': '<script>' });
+        ssrContainer.addRoot({ '<scrip></script>': '"<scrip></script>"', '<': '<script>' });
       });
       expect(container.getObjectById(0)).toEqual({
         '<scrip></script>': '"<scrip></script>"',
@@ -95,11 +96,11 @@ describe('serializer v2', () => {
     it('should serialize non-standard objects', () => {
       const container = withContainer((ssrContainer) => {
         const obj = { null: null, undefined: undefined };
-        expect(ssrContainer.getObjectId(null)).toBe(0);
-        expect(ssrContainer.getObjectId(undefined)).toBe(1);
-        expect(ssrContainer.getObjectId(obj)).toBe(2);
+        expect(ssrContainer.addRoot(null)).toBe(0);
+        expect(ssrContainer.addRoot(undefined)).toBe(1);
+        expect(ssrContainer.addRoot(obj)).toBe(2);
         expect(
-          ssrContainer.getObjectId([null, undefined, obj, { null: null, undefined: undefined }])
+          ssrContainer.addRoot([null, undefined, obj, { null: null, undefined: undefined }])
         ).toBe(3);
       });
       const obj = container.getObjectById(2);
@@ -118,8 +119,8 @@ describe('serializer v2', () => {
     it('should de-dup long strings', () => {
       const str = new Array(100).fill('a').join('');
       const container = withContainer((ssrContainer) => {
-        expect(ssrContainer.getObjectId(str)).toBe(0);
-        expect(ssrContainer.getObjectId({ a: str, b: str })).toBe(1);
+        expect(ssrContainer.addRoot(str)).toBe(0);
+        expect(ssrContainer.addRoot({ a: str, b: str })).toBe(1);
       });
       const idx = container.element.innerHTML.indexOf(str);
       expect(idx).toBeGreaterThan(0);
@@ -137,7 +138,7 @@ describe('serializer v2', () => {
           qrl('chunk.js', 's_123', ['Hello', 'World']) as QRLInternal,
           inlinedQrl(testFn, 's_inline') as QRLInternal,
         ];
-        const [qrl0, qrl1, qrl2] = withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0);
+        const [qrl0, qrl1, qrl2] = withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0);
         expect(qrl0.$hash$).toEqual(obj[0].$hash$);
         expect(qrl0.$captureRef$).toEqual(obj[0].$captureRef$);
         expect(qrl0._devOnlySymbolRef).toEqual((obj[0] as any)._devOnlySymbolRef);
@@ -162,38 +163,38 @@ describe('serializer v2', () => {
     describe('URLSerializer, ////////////// \u0005', () => {
       it('should serialize and deserialize', () => {
         const obj = new URL('http://server/path#hash');
-        expect(withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0)).toEqual(obj);
+        expect(withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0)).toEqual(obj);
       });
     });
     describe('DateSerializer, ///////////// \u0006', () => {
       it('should serialize and deserialize', () => {
         const obj = new Date();
-        expect(withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0)).toEqual(obj);
+        expect(withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0)).toEqual(obj);
       });
     });
     describe('RegexSerializer, //////////// \u0007', () => {
       it('should serialize and deserialize', () => {
         const obj = /abc/gim;
-        expect(withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0)).toEqual(obj);
+        expect(withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0)).toEqual(obj);
       });
     });
     describe('ErrorSerializer, //////////// \u000E', () => {
       it('should serialize and deserialize', () => {
         const obj = Object.assign(new Error('MyError'), { extra: 'property' });
-        expect(withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0)).toEqual(obj);
+        expect(withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0)).toEqual(obj);
       });
     });
     describe('DocumentSerializer, ///////// \u000F', () => {
       it('should serialize and deserialize', () => {
         const obj = new SsrNode(SsrNode.DOCUMENT_NODE);
-        const container = withContainer((ssr) => ssr.getObjectId(obj));
+        const container = withContainer((ssr) => ssr.addRoot(obj));
         expect(container.getObjectById(0)).toEqual(container.element.ownerDocument);
       });
     });
     describe('ComponentSerializer, //////// \u0010', () => {
       it('should serialize and deserialize', () => {
         const obj = component$(() => <div />);
-        const container = withContainer((ssr) => ssr.getObjectId(obj));
+        const container = withContainer((ssr) => ssr.addRoot(obj));
         const [srcQrl] = (obj as any)[SERIALIZABLE_STATE];
         const [dstQrl] = container.getObjectById(0)[SERIALIZABLE_STATE];
         expect(dstQrl.$hash$).toEqual(srcQrl.$hash$);
@@ -219,13 +220,13 @@ describe('serializer v2', () => {
     describe('NaN, //////////////////////// \u0014', () => {
       it('should serialize and deserialize', () => {
         const obj = Number.NaN;
-        expect(withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0)).toEqual(obj);
+        expect(withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0)).toEqual(obj);
       });
     });
     describe('URLSearchParamsSerializer, // \u0015', () => {
       it('should serialize and deserialize', () => {
         const obj = new URLSearchParams('a=1&b=2');
-        expect(withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0)).toEqual(obj);
+        expect(withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0)).toEqual(obj);
       });
     });
     describe('FormDataSerializer, ///////// \u0016', () => {
@@ -241,20 +242,20 @@ describe('serializer v2', () => {
     describe('BigIntSerializer, /////////// \u0018', () => {
       it('should serialize and deserialize', () => {
         const obj = BigInt('12345678901234567890');
-        expect(withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0)).toEqual(obj);
+        expect(withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0)).toEqual(obj);
       });
     });
     describe('SetSerializer, ////////////// \u0019', () => {
       it('should serialize and deserialize', () => {
         const obj = new Set(['a', 'b', 'c']);
-        expect(withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0)).toEqual(obj);
+        expect(withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0)).toEqual(obj);
       });
       it('should dedup internal state', () => {
         const a = { a: 'A' };
         const b = { b: 'B', a: a };
         const c = { c: 'C', a: a, b: b };
         const obj = new Set([a, b, c]);
-        const value: Set<any> = withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0);
+        const value: Set<any> = withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0);
         expect(value).toEqual(obj);
         const [valueA, valueB, valueC] = Array.from(value.values());
         expect(valueB.a).toBe(valueA);
@@ -269,7 +270,7 @@ describe('serializer v2', () => {
           ['b', 3],
           ['c', 4],
         ]);
-        expect(withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0)).toEqual(obj);
+        expect(withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0)).toEqual(obj);
       });
       it('should dedup internal state', () => {
         const a = { a: 'A' };
@@ -280,9 +281,7 @@ describe('serializer v2', () => {
           ['b', b],
           ['c', c],
         ]);
-        const value: Map<string, any> = withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(
-          0
-        );
+        const value: Map<string, any> = withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0);
         expect(value).toEqual(obj);
         const [valueA, valueB, valueC] = Array.from(value.values());
         expect(valueB.a).toBe(valueA);
@@ -292,14 +291,14 @@ describe('serializer v2', () => {
     });
     describe('StringSerializer, /////////// \u001b', () => {
       it('should serialize and deserialize', () => {
-        const obj = SerializationConstant.Document_CHAR + 'anything';
-        expect(withContainer((ssr) => ssr.getObjectId(obj)).getObjectById(0)).toEqual(obj);
+        const obj = '\u0010anything';
+        expect(withContainer((ssr) => ssr.addRoot(obj)).getObjectById(0)).toEqual(obj);
       });
     });
   });
 });
 
-function withContainer(ssrFn: (ssrContainer: any) => void): Container {
+function withContainer(ssrFn: (ssrContainer: SSRContainer) => void): Container {
   const ssrContainer = ssrCreateContainer();
   ssrContainer.openContainer();
   ssrFn(ssrContainer);
